+++ src/MrCatzEvent.php
@@ -25,6 +25,7 @@ final class MrCatzEvent
     public const BULK_DELETE = 'bulkDeleteData';
     public const BULK_ACTION = 'bulkActionData';
     public const BULK_ACTION_OPEN = 'bulkActionOpen';
+    public const BULK_ACTION_DONE = 'bulkActionDone';
     public const REFRESH_TABLE = 'refreshDataTable';
     public const INLINE_UPDATE = 'inlineUpdateData';
     public const ROW_CLICK = 'rowClickData';